+++ src/vue/modules/create-sale-form/mixins/manage-sale-request.mixin.spec.js
@@ -101,8 +101,8 @@ describe('Manage sale request mixin', () => {
           .to.equal('10.000000')
         expect(wrapper.vm.saleRequestOpts.quoteAssets)
           .to.deep.equal([
-            { asset: 'BTC', price: '1' },
-            { asset: 'ETH', price: '1' },
+            { asset: 'BTC', price: '0.050000' },
+            { asset: 'ETH', price: '0.050000' },
           ])
 
         expect(wrapper.vm.saleRequestOpts.creatorDetails.name)
@@ -180,6 +180,7 @@ describe('Manage sale request mixin', () => {
 
           sandbox.stub(DocumentUploader, 'uploadDocument').resolves()
           sandbox.stub(wrapper.vm, 'createBalancesIfNotExist').resolves()
+          sandbox.stub(wrapper.vm, 'loadAssetsPairsByQuote').resolves([])
 
           sandbox.stub(wrapper.vm, 'createSaleDescriptionBlob')
             .resolves('BLOB_ID')